refactored tables in orders and cart. Also created firebaseController.js to concentrate all firebase related functions
