+++ src/components/Item/Item.js
@@ -8,14 +8,13 @@ import { Link } from 'react-router-dom';
 const Item = ({ product }) => {
     const { name, price, image, stock, id } = product;
     return (
-        //JSX
         <Card>
             <CardContent>
                 <div className="cardItem">
                     <Link style={{ textDecoration: 'none', color: 'inherit' }} to={`/item/${id}`}>
                         <img src={`${image}`} alt={name} />
+                        <p>{name}</p>
                     </Link>
-                    <p>{name}</p>
                     <small>Stock: {stock} u.</small>
                     <span>$ {price}</span>
                     <ItemCount product={product} />
