+++ src/components/OrdersContainer/OrdersContainer.js
@@ -1,16 +1,21 @@
+// React imports
 import { useEffect, useState } from "react"
+// Site components imports
+import LoadingSpinner from "../LoadingSpinner/LoadingSpinner"
+import Orders from "../Orders/Orders"
+// Firebase imports
 import db from "../../utils/firebaseConfig"
-import { collection, getDocs } from "firebase/firestore";
-import Orders from "../Orders/Orders";
-import LoadingSpinner from "../LoadingSpinner/LoadingSpinner";
+import { collection, getDocs } from "firebase/firestore"
+
+
 
 const OrdersContainer = () => {
-const [orders, setOrders] = useState([]);
-const [spinnerState, setSpinnerState] = useState({display: 'flex'})
+    const [orders, setOrders] = useState([]);
+    const [spinnerState, setSpinnerState] = useState({ display: 'flex' })
 
     useEffect(() => {
         getOrdersFromFirestore()
-        .then(() => setSpinnerState({display: 'none'}))
+            .then(() => setSpinnerState({ display: 'none' }))
     }, [])
 
     const getOrdersFromFirestore = async () => {
@@ -20,17 +25,17 @@ const [spinnerState, setSpinnerState] = useState({display: 'flex'})
             order.id = doc.id;
             return order;
         })
-        
+
         setOrders(orderList);
     }
 
     return (
         <>
-        <h1>Mis Órdenes</h1>
-        {(orders.length > 0) && <Orders orders={orders} /> }
-        <LoadingSpinner display={spinnerState} />
+            <h1>Mis Órdenes</h1>
+            {(orders.length > 0) && <Orders orders={orders} />}
+            <LoadingSpinner display={spinnerState} />
         </>
-        
+
     )
 }
